+++ src/main/java/yoon/community/repository/board/BoardRepository.java
@@ -9,11 +9,11 @@ import yoon.community.entity.user.User;
 
 import java.util.List;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
 
     Page<Board> findByTitleContaining(String keyword, Pageable pageable);
     Page<Board> findAll(Pageable pageable);
-    Page<Board> findAllByCategoryId(Pageable pageable, int id);
+    Page<Board> findAllByCategoryId(Pageable pageable, Long id);
     Page<Board> findByLikedGreaterThanEqual(Pageable pageable, int number);
     List<Board> findByReportedIsTrue();
 }
