+++ src/main/java/yoon/community/repository/board/FavoriteRepository.java
@@ -8,7 +8,7 @@ import yoon.community.entity.user.User;
 
 import java.util.List;
 
-public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
+public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
     Optional<Favorite> findFavoriteByBoard(Board board);
     Optional<Favorite> findByBoardAndUser(Board board, User user);
     List<Favorite> findAllByUser(User user);
